# docs/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API сервиса мониторинга процесса эмульсификации (партии, рецепты, пользователи).",
        "title": "EBR Monitoring Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "description": "Вход по username/password, возвращает JWT-токен",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Авторизация",
                "parameters": [
                    {
                        "description": "учётные данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "invalid login or password",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "account disabled",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/batches": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batches"
                ],
                "summary": "Список партий по статусу",
                "parameters": [
                    {
                        "type": "string",
                        "description": "статус партии",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.GetBatchesByStatusResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "query parameter is required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to list batches",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт партию по рецепту. registered_by берётся из JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batches"
                ],
                "summary": "Создать партию (batch)",
                "parameters": [
                    {
                        "description": "данные партии",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.CreateBatchResponse"
                        }
                    },
                    "400": {
                        "description": "invalid json or invalid batch volume",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "recipe not found or archived",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to create batch",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/recipes/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Получить рецепт по коду",
                "parameters": [
                    {
                        "type": "string",
                        "description": "код рецепта",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.GetRecipeByCodeResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "recipe not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "recipe archived",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/users": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт нового пользователя (роль admin или operator). Доступно только админу.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Создать пользователя",
                "parameters": [
                    {
                        "description": "данные пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.CreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "validation error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "forbidden / invalid role",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "user already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.CreateBatchRequest": {
            "type": "object",
            "properties": {
                "recipe_code": {
                    "type": "string"
                },
                "target_volume_l": {
                    "type": "integer"
                }
            }
        },
        "domain.CreateBatchResponse": {
            "type": "object",
            "properties": {
                "batch_code": {
                    "type": "string"
                },
                "batch_status": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "registered_by": {
                    "type": "integer"
                }
            }
        },
        "domain.CreateUserRequest": {
            "type": "object",
            "properties": {
                "father_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "domain.CreateUserResponse": {
            "type": "object",
            "properties": {
                "user_code": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "domain.GetBatchesByStatusResponse": {
            "type": "object",
            "properties": {
                "batch_code": {
                    "type": "string"
                },
                "batch_id": {
                    "type": "integer"
                },
                "batch_status": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "recipe_code": {
                    "type": "string"
                },
                "registered_by": {
                    "type": "string"
                },
                "target_volume_l": {
                    "type": "integer"
                }
            }
        },
        "domain.GetRecipeByCodeResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "max_volume_l": {
                    "type": "integer"
                },
                "min_volume_l": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "required_equipment_type": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "domain.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "domain.LoginResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user_code": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT токен в формате: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}